fix(video-still-service): set correct video still when cutting videos
